Rename lineplot to line
